Change to non dependent plugin
